Moving the TrustDomain and DeviceGroup out of the cluster/__init__.py and up into f5/multi_device.
